refactor: socketController 리팩토링
