+++ controllers/actionSocketController.js
@@ -6,27 +6,25 @@ const utilities = require('../modules/utility');
 
 module.exports = function (io) {
   io.on('connection', function (socket) {
-    socket.on('enterLobby', enterLobby);
-    socket.on('createRoom', createRoom);
-    socket.on('joinRoom', joinRoom);
-    socket.on('exitRoom', exitRoom);
-    socket.on('startRound', startRound);
-    socket.on('endRound', endRound);
-
-    socket.on('endCycleHarvestCrop', endCycleHarvestCrop);
-    socket.on('endCyclePayFood', endCyclePayFood);
-    socket.on('endCycleBreedAnimal', endCycleBreedAnimal);
-
-    socket.on('startGame', startGame);
-    socket.on('useFacility', useFacility);
-
-    socket.on('useActionSpace', useActionSpace);
     socket.on('endTurn', endTurn);
-
-    socket.on('endGame', endGame);
-    socket.on('skipGame', skipGame);
+    socket.on('useFacility', useFacility);
+    socket.on('useActionSpace', useActionSpace);
     socket.on('accumulateGoods', accumulateGoods);
     socket.on('transformData', transformData);
+
+    async function endTurn(data) {
+      let userId = data.userId;
+      let roomId = data.roomId;
+
+      let updatedStatus = await gameService.endTurn(roomId, userId);
+      if (updatedStatus.isSuccess === false) {
+        io.sockets.emit('endTurn', baseResponse.BAD_REQUEST);
+        return;
+      } else {
+        io.sockets.emit('endTurn', updatedStatus);
+        return;
+      }
+    }
 
     async function transformData(data) {
       socket.emit('transformData', data);
@@ -162,59 +160,6 @@ module.exports = function (io) {
         );
         io.sockets.emit('useFacility', updatedPlayer);
       }
-    }
-
-    async function endTurn(data) {
-      let userId = data.userId;
-      let roomId = data.roomId;
-
-      let updatedStatus = await gameService.endTurn(roomId, userId);
-      if (updatedStatus.isSuccess === false) {
-        io.sockets.emit('endTurn', baseResponse.BAD_REQUEST);
-        return;
-      } else {
-        io.sockets.emit('endTurn', updatedStatus);
-        return;
-      }
-    }
-
-    async function startGame(data) {
-      // data 형식
-      // [
-      //   {
-      //     roomId: 1,
-      //     userId: 'test1',
-      //   },
-      //   {
-      //     roomId: 1,
-      //     userId: 'test2',
-      //   },
-      // ];
-      for (let roomData of data) {
-        let roomId = roomData.roomId;
-        let userId = roomData.userId;
-        let isStart = await roomService.checkIsInGameStatus(roomId, userId);
-        if (isStart) {
-          io.sockets.emit('startGame', errResponse(baseResponse.BAD_REQUEST));
-          return;
-        } else {
-          await gameService.startGame(roomId, userId);
-        }
-      }
-      let gameStatus = await roomService.getGameStatus(data[0].roomId);
-      let mainFacilityList = await gameService.getMainFacilityCards(
-        data[0].roomId
-      );
-      io.sockets.emit(
-        'startGame',
-        response(baseResponse.SUCCESS, {
-          gameStatusList: gameStatus,
-          mainFacilityList: mainFacilityList,
-        })
-      );
-
-      let updatedRooms = await roomService.getRooms();
-      io.sockets.emit('updatedRooms', updatedRooms);
     }
 
     async function useActionSpace(data) {
@@ -629,321 +574,6 @@ module.exports = function (io) {
         );
         io.sockets.emit('useActionSpace', updateResult);
       }
-    }
-
-    async function exitRoom(data) {
-      try {
-        let userId = data.userId;
-        let roomId = data.roomId;
-
-        let isHost = await roomService.checkIsHost(roomId, userId);
-
-        if (isHost) {
-          // Delete the room
-          await roomService.deleteRoom(roomId);
-          let updatedRooms = await roomService.getRooms();
-          io.sockets.emit('updatedRooms', updatedRooms);
-          return;
-        }
-
-        // Check if the user is in the room
-        let isInRoom = await roomService.checkIsInRoom(roomId, userId);
-        if (!isInRoom) {
-          io.sockets.emit(
-            'exitRoom',
-            errResponse(baseResponse.ROOM_NOT_JOINED)
-          );
-          return;
-        }
-
-        // Subtract the participant number
-        await roomService.calParticipantNum(roomId, false);
-
-        // Delete the user from the room
-        await roomService.exitRoom(roomId, userId);
-
-        let updatedRooms = await roomService.getRooms();
-        io.sockets.emit('updatedRooms', updatedRooms);
-        // TODO: 방 안에도 보내주는 emit 필요
-        let playerInRoom = await roomService.findUserListByRoomId(roomId);
-        io.sockets.emit('exitRoom', playerInRoom);
-      } catch (err) {
-        console.log(err);
-        io.sockets.emit('exitRooms', errResponse(baseResponse.SERVER_ERROR));
-      }
-    }
-
-    async function enterLobby() {
-      try {
-        let updatedRooms = await roomService.getRooms();
-        io.sockets.emit('updatedRooms', updatedRooms);
-      } catch (err) {
-        console.log(err);
-        io.sockets.emit(
-          'patchRoomList',
-          errResponse(baseResponse.SERVER_ERROR)
-        );
-      }
-    }
-
-    async function createRoom(data) {
-      try {
-        let roomName = data.roomName;
-        let limitNum = data.limitNum;
-        let hostId = data.hostId;
-
-        // Check if the host is already in the room
-        let isInRoom = await roomService.isInRoom(hostId);
-        if (isInRoom) {
-          io.sockets.emit(
-            'createRoom',
-            errResponse(baseResponse.ALREADY_IN_ROOM)
-          );
-        } else {
-          // Create the room
-          let createResult = await roomService.createRoom(
-            roomName,
-            limitNum,
-            hostId
-          );
-
-          if (createResult.isSuccess === false) {
-            io.sockets.emit('createRoom', createResult);
-          } else {
-            // Find the room id
-            let roomId = await roomService.findRoomId(hostId);
-
-            // Add the host to the room
-            await roomService.joinRoom(
-              parseInt(roomId.dataValues.room_id),
-              hostId
-            );
-            socket.join(roomId.dataValues.room_id);
-            let updatedRooms = await roomService.getRooms();
-            io.sockets.emit('updatedRooms', updatedRooms);
-          }
-        }
-      } catch (err) {
-        console.log(err);
-        io.sockets.emit('createRoom', errResponse(baseResponse.SERVER_ERROR));
-      }
-    }
-
-    async function joinRoom(data) {
-      try {
-        let roomId = data.roomId;
-        let userId = data.userId;
-
-        // Check if the user is already in the room
-        let isInRoom = await roomService.isInRoom(userId);
-        if (isInRoom) {
-          io.sockets.emit(
-            'joinRoom',
-            errResponse(baseResponse.ALREADY_IN_ROOM)
-          );
-        } else {
-          // Add the participant number
-          let calResult = await roomService.calParticipantNum(roomId, true);
-
-          if (calResult.isSuccess === false) {
-            io.sockets.emit('joinRoom', calResult);
-          } else {
-            // Add the user to the room
-            await roomService.joinRoom(roomId, userId);
-            socket.join(parseInt(roomId));
-
-            let updatedRooms = await roomService.getRooms();
-            io.sockets.emit('updatedRooms', updatedRooms);
-            // TODO: 방 안에도 보내주는 emit 필요
-            let playerInRoom = await roomService.findUserListByRoomId(roomId);
-            io.sockets.emit('joinRoom', playerInRoom);
-          }
-        }
-      } catch (err) {
-        console.log(err);
-        io.sockets.emit('joinRoom', errResponse(baseResponse.SERVER_ERROR));
-      }
-    }
-
-    async function endCycleHarvestCrop(data) {
-      try {
-        let roomId = data.roomId;
-        let userIdList = await roomService.findUserListByRoomId(roomId);
-        // 작물 수확
-        let result = await gameService.harvestCrop(userIdList, roomId);
-        if (result.isSuccess === false) {
-          io.sockets.emit('endCycle', result);
-          return;
-        }
-        io.sockets.emit(
-          'endCycleHarvestCrop',
-          response(baseResponse.SUCCESS, result)
-        );
-      } catch (err) {
-        console.log(err);
-        io.sockets.emit(
-          'endCycleHarvestCrop',
-          errResponse(baseResponse.SERVER_ERROR)
-        );
-      }
-    }
-
-    async function endCyclePayFood(data) {
-      try {
-        let roomId = data.roomId;
-        let userIdList = await roomService.findUserListByRoomId(roomId);
-        console.log(userIdList);
-        //음식 지불
-        let result = await gameService.payFood(userIdList, roomId);
-        if (result.isSuccess === false) {
-          io.sockets.emit('endCycle', result);
-          return;
-        }
-        io.sockets.emit(
-          'endCyclePayFood',
-          response(baseResponse.SUCCESS, result)
-        );
-      } catch (err) {
-        console.log(err);
-        io.sockets.emit(
-          'endCyclePayFood',
-          errResponse(baseResponse.SERVER_ERROR)
-        );
-      }
-    }
-
-    async function endCycleBreedAnimal(data) {
-      try {
-        let roomId = data.roomId;
-        let userIdList = await roomService.findUserListByRoomId(roomId);
-
-        //음식 지불
-        let result = await gameService.breedAnimal(userIdList, roomId);
-        if (result.isSuccess === false) {
-          io.sockets.emit('endCycle', result);
-          return;
-        }
-        io.sockets.emit(
-          'endCycleBreedAnimal',
-          response(baseResponse.SUCCESS, result)
-        );
-      } catch (err) {
-        console.log(err);
-        io.sockets.emit(
-          'endCycleBreedAnimal',
-          errResponse(baseResponse.SERVER_ERROR)
-        );
-      }
-    }
-
-    async function startRound(data) {
-      let roomId = data.roomId;
-      let findUserList = await roomService.findUserListByRoomId(roomId);
-      if (findUserList.length == 0) {
-        io.sockets.emit('startRound', baseResponse.BAD_REQUEST);
-        return;
-      }
-      // 아기 성장
-      let growBabyResult = await gameService.growBaby(findUserList, roomId);
-      if (growBabyResult.isSuccess == false) {
-        io.sockets.emit('startRound', growBabyResult);
-        return;
-      }
-
-      // 게임 순서 변경
-      let updateOrderResult = await gameService.updateOrder(
-        findUserList,
-        roomId
-      );
-      if (updateOrderResult.isSuccess == false) {
-        io.sockets.emit('startRound', baseResponse.BAD_REQUEST);
-        return;
-      }
-      // 구성물 수거 (소규모 농부, Small farmer)
-      let usedJobCardResult;
-      let updateResult = [];
-      for (let userId of findUserList) {
-        usedJobCardResult = await gameService.usedJobCard(
-          userId,
-          roomId,
-          'Small farmer'
-        );
-        if (usedJobCardResult.isSuccess == false) {
-          io.sockets.emit('startRound', baseResponse.BAD_REQUEST);
-          return;
-        }
-
-        updateResult.push(await gameService.getPlayerStatus(userId, roomId));
-      }
-      io.sockets.emit('startRound', {
-        updateOrderResult,
-        updateResult,
-      });
-      // io.sockets.emit('startRound', {
-      //   updateOrderResult,
-      //   'Small Farmer': updateResult,
-      // });
-    }
-
-    async function endRound(data) {
-      io.sockets.emit('endRound', 'endRound');
-    }
-
-    async function endGame(data) {
-      let roomId = data.roomId;
-      let calGameScoreResult = await gameService.calGameScore(roomId);
-
-      io.sockets.emit('endGame', calGameScoreResult);
-    }
-
-    async function skipGame(data) {
-      // data 형식
-      // {
-      //   "roomId": 1,
-      //   "skipRound": 8,
-      //   "userId": [
-      //       {"userId": "test1"},
-      //       {"userId": "test2"}
-      //     ]
-      //  }
-      // 시나리오 기준, P1(빨)이 무조건 userId List의 첫번째에 위치해야 함
-
-      roomId = data.roomId;
-      let i = 0;
-      let skipResult = [];
-      let accumulateGoods = [];
-      for (let obj of data.userId) {
-        try {
-          let userId = obj.userId;
-          let result = await gameService.skipGame(
-            data.roomId,
-            userId,
-            data.skipRound,
-            i
-          );
-          i++;
-          if (result.isSuccess === false) {
-            io.sockets.emit('skipGame', result);
-          } else {
-            let updatedPlayer = await gameService.getPlayerStatus(
-              userId,
-              roomId
-            );
-            skipResult.push(updatedPlayer);
-            accumulateGoods = await gameService.getAccumulatedGoodsByRoomId(
-              roomId
-            );
-          }
-        } catch (err) {
-          console.log(err);
-          io.sockets.emit('skipGame', errResponse(baseResponse.SERVER_ERROR));
-        }
-      }
-      io.sockets.emit('skipGame', {
-        updatedPlayer: skipResult,
-        accResult: accumulateGoods,
-        skipRound: data.skipRound,
-      });
     }
 
     async function accumulateGoods(data) {
